% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fat_models}
\alias{fat_models}
\title{Reference Models for Body Composition}
\format{
\subsection{\code{fat_models}}{

A named list of adipose tissue models. Naming scheme: (metric)\emph{(sex)}(vertebral level)
}
}
\source{
\url{https://bodycomp-metrics.mgh.harvard.edu}
}
\usage{
fat_models
}
\description{
LMSP models for age- and sex-specific reference values for CT-assessed adipose tissue.
}
\keyword{datasets}
